Split LibProfiler into smaller libraries
