+++ src/SharpDetect.Profiler/LibIL/ILRewriter.cpp
