+++ src/SharpDetect.Profiler/LibIL/ILRewriter.h
