+++ src/SharpDetect.Profiler/LibIL/MetadataAnalysis.cpp
