+++ src/SharpDetect.Profiler/LibIL/MetadataAnalysis.h
