+++ src/SharpDetect.Profiler/LibMetadata/AssemblyDef.cpp
