+++ src/SharpDetect.Profiler/LibMetadata/AssemblyDef.h
