+++ src/SharpDetect.Profiler/LibMetadata/ModuleDef.cpp
