+++ src/SharpDetect.Profiler/LibMetadata/ModuleDef.h
