+++ src/SharpDetect.Profiler/LibMetadata/SignatureUtils.cpp
