+++ src/SharpDetect.Profiler/LibMetadata/SignatureUtils.h
