+++ src/SharpDetect.Profiler/LibMetadata/WString.cpp
