+++ src/SharpDetect.Profiler/LibMetadata/WString.h
