+++ src/SharpDetect.Profiler/LibProfilerCore/ClassFactory.h
