+++ src/SharpDetect.Profiler/LibProfilerCore/ComPtr.h
