+++ src/SharpDetect.Profiler/LibProfilerCore/CorProfilerBase.cpp
