+++ src/SharpDetect.Profiler/LibProfilerCore/CorProfilerBase.h
