+++ src/SharpDetect.Profiler/LibProfilerCore/GarbageCollectionContext.cpp
