+++ src/SharpDetect.Profiler/LibProfilerCore/GarbageCollectionContext.h
