+++ src/SharpDetect.Profiler/LibProfilerCore/ObjectsTracker.cpp
