+++ src/SharpDetect.Profiler/LibProfilerCore/ObjectsTracker.h
