+++ src/SharpDetect.Profiler/LibProfilerCore/PAL.cpp
