+++ src/SharpDetect.Profiler/LibProfilerCore/PAL.h
