+++ src/SharpDetect.Profiler/LibProfilerCore/StackWalker.cpp
