+++ src/SharpDetect.Profiler/LibProfilerCore/StackWalker.h
