+++ src/SharpDetect.Profiler/LibProfilerCore/TrackedObjectId.h
