+++ src/SharpDetect.Profiler/LibProfilerCore/asmHelpers.S
