+++ src/SharpDetect.Profiler/LibProfilerCore/asmHelpers.asm
