+++ src/SharpDetect.Profiler/SharpDetect.Concurrency.Profiler/CMakeLists.txt
@@ -11,7 +11,7 @@ set(SOURCES
     "TypeInjectionDescriptor.cpp")
 
 add_library (SharpDetect.Concurrency.Profiler SHARED ${SOURCES} "Profiler.def")
-target_link_libraries (SharpDetect.Concurrency.Profiler PRIVATE LibProfiler LibIPC)
+target_link_libraries (SharpDetect.Concurrency.Profiler PRIVATE LibProfilerCore LibMetadata LibIL LibIPC)
 apply_profiler_compile_options(SharpDetect.Concurrency.Profiler)
 if(UNIX)
     target_link_options(SharpDetect.Concurrency.Profiler PRIVATE "LINKER:-z,noexecstack")
